DeepJanus-MNIST/requirements.txt
numpy
h5py
potracer
cairosvg
pillow
deap
matplotlib
tensorflow==2.15.0
